# src/main/resources/openapi/todo-rest-api-ext.yaml
openapi: 3.0.1
info:
  title: ToDo API
  description: API for managing ToDo tasks
  version: 1.0.0

servers:
  - url: http://localhost:8080

paths:
  /todos:
    get:
      summary: List all ToDos (Paged)
      operationId: fetchTodosPaged
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 0
          required: false
          description: Page number
        - in: query
          name: size
          schema:
            type: integer
            default: 10
          required: false
          description: Page size
      responses:
        '200':
          description: A paged list of ToDos
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedToDoResponse'
    post:
      summary: Create a new ToDo
      operationId: createTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToDo'
      responses:
        '201':
          description: Created
    put:
      summary: Update a ToDo
      operationId: updateTodo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ToDo'
      responses:
        '200':
          description: Updated
  /todos/{id}:
    get:
      summary: Get a ToDo by ID
      operationId: getTodo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: A single ToDo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToDo'
    delete:
      summary: Delete a ToDo
      operationId: deleteTodo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: No Content

components:
  schemas:
    ToDo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        version:
          type: integer
        title:
          type: string
        description:
          type: string
        grp:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date
        status:
          type: string
          enum:
            - PENDING
            - ONGOING
            - DISMISSED
            - DONE
      required:
        - title
        - status

    PagedToDoResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ToDo'
        totalElements:
          type: integer
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer